extract opacity component
